Add the region populations for the 12 RCI groups
We only provide the totals for the individual RCI groups because the
game already provides totals for the overall RCI population.

The game's overall RCI population values may come from a different
source than our totals for the individual groups, but that should be
fine as the overall total rarely if ever equals the sum of the
individual group totals.